requirements.txt
openai>=1.0
gradio>=4.0
python-dotenv>=1.0
